+++ example/lib/main.dart
@@ -67,7 +67,7 @@ class MyHomePage extends StatelessWidget {
                 context,
                 blur: 0,
                 dialogTransitionType: DialogTransitionType.Shrink,
-                transitionDuration: Duration(milliseconds: 100),
+                // transitionDuration: Duration(milliseconds: 100),
                 onDismiss: () {
                   print("Do something onDismiss");
                 },
@@ -75,7 +75,8 @@ class MyHomePage extends StatelessWidget {
               progressDialog.setLoadingWidget(CircularProgressIndicator(
                 valueColor: AlwaysStoppedAnimation(Colors.red),
               ));
-              progressDialog.setMessage(Text("Please Wait, Injecting your phone with my virus"));
+              progressDialog.setMessage(
+                  Text("Please Wait, Injecting your phone with my virus"));
               progressDialog.setTitle(Text("Loading"));
               progressDialog.show();
 
@@ -89,7 +90,8 @@ class MyHomePage extends StatelessWidget {
 
               progressDialog.dismiss();
             },
-            child: Text("Progress Dialog", style: TextStyle(color: Colors.white)),
+            child:
+                Text("Progress Dialog", style: TextStyle(color: Colors.white)),
           ),
           FlatButton(
             color: Colors.blue,
@@ -113,7 +115,8 @@ class MyHomePage extends StatelessWidget {
 
               progressDialog.dismiss();
             },
-            child: Text("Custom Progress Dialog", style: TextStyle(color: Colors.white)),
+            child: Text("Custom Progress Dialog",
+                style: TextStyle(color: Colors.white)),
           ),
           FlatButton(
             color: Colors.blue,
@@ -138,7 +141,8 @@ class MyHomePage extends StatelessWidget {
                 title: Text("Loging in"),
               ));
             },
-            child: Text("Progress Dialog Future", style: TextStyle(color: Colors.white)),
+            child: Text("Progress Dialog Future",
+                style: TextStyle(color: Colors.white)),
           ),
           FlatButton(
             color: Colors.blue,
@@ -160,7 +164,8 @@ class MyHomePage extends StatelessWidget {
                 },
               ));
             },
-            child: Text("Custom Progress Dialog Future", style: TextStyle(color: Colors.white)),
+            child: Text("Custom Progress Dialog Future",
+                style: TextStyle(color: Colors.white)),
           ),
           FlatButton(
             color: Colors.blue,
@@ -185,7 +190,8 @@ class MyHomePage extends StatelessWidget {
                 ],
               ).show(context);
             },
-            child: Text("Show NAlertDialog", style: TextStyle(color: Colors.white)),
+            child: Text("Show NAlertDialog",
+                style: TextStyle(color: Colors.white)),
           ),
           FlatButton(
             color: Colors.blue,
@@ -228,7 +234,8 @@ class MyHomePage extends StatelessWidget {
                 blur: 0,
                 dialog: AlertDialog(
                   title: Text("Alert Dialog"),
-                  content: Text("Wohoo.. This is ordinary AlertDialog with Blur background"),
+                  content: Text(
+                      "Wohoo.. This is ordinary AlertDialog with Blur background"),
                   actions: <Widget>[
                     FlatButton(
                       child: Text("You"),
@@ -265,7 +272,8 @@ class MyHomePage extends StatelessWidget {
                 backgroundColor: Colors.red.withOpacity(.5),
                 blur: 0,
                 title: Text("Alert Dialog"),
-                content: Text("Wohoo.. This is ordinary AlertDialog with Custom Color background"),
+                content: Text(
+                    "Wohoo.. This is ordinary AlertDialog with Custom Color background"),
                 actions: <Widget>[
                   FlatButton(
                     child: Text("You"),
@@ -294,7 +302,8 @@ class MyHomePage extends StatelessWidget {
             onPressed: () async {
               await AlertDialog(
                 title: Text("Alert Dialog"),
-                content: Text("Wohoo.. This is ordinary AlertDialog with Blur background"),
+                content: Text(
+                    "Wohoo.. This is ordinary AlertDialog with Blur background"),
                 actions: <Widget>[
                   FlatButton(
                     child: Text("You"),
@@ -323,7 +332,8 @@ class MyHomePage extends StatelessWidget {
             onPressed: () async {
               await CupertinoAlertDialog(
                 title: Text("Alert Dialog"),
-                content: Text("Wohoo.. This is ordinary AlertDialog with Blur background"),
+                content: Text(
+                    "Wohoo.. This is ordinary AlertDialog with Blur background"),
                 actions: <Widget>[
                   FlatButton(
                     child: Text("You"),